refactor(daemon): host the deferred-outcome seam in the stabilization owner (#1633 review)
Zero R9 growth, per review: a NEW aggregator file cannot stay out of the
cycle by shedding type imports — its value imports of the two member owners
close the loop regardless. The unique zero-growth host is an existing cycle
node, and the stabilization owner is the only legal one (the policy module
cannot value-import stabilization back, R4; the freshness module would join
as a new member). So deferred-interaction-outcome.ts absorbs the
post-gesture-stabilization implementation and becomes the R7 owner of
postGestureStabilization: the seam lives in a node that was already on the
member-to-member paths it concentrates.

- markPostGestureStabilization is now module-private behind
  markDeferredInteractionOutcome; stabilization marking tests drive the
  public interface.
- The freshness retry loop moves beside its classifier in
  android-snapshot-freshness.ts; gesture-no-effect helpers move to a leaf
  file. Both stay outside the cycle.
- Ratchet reverted to main's exact values (76 files, daemon-server 20) —
  measured member-by-member vs main: the diff is empty.
