+++ CONTEXT.md
@@ -118,7 +118,8 @@ task touches:
   take effect" — pending interaction outcome retry, post-gesture stabilization, and Android
   snapshot freshness recovery. `src/daemon/deferred-interaction-outcome.ts` is its one interface:
   every mutating route marks through it after dispatch, and every snapshot capture resolves
-  through it; the three `SessionState` fields and their owner modules are implementation. Marking
+  through it; the three `SessionState` fields stay with their R7 owners (the module itself owns
+  `postGestureStabilization`, so the seam adds no node to the R9 cycle). Marking
   order is load-bearing (pending outcome retry before stabilization), each marker keeps its own
   eligibility gate, and the module owns only these post-action markers — never ADR 0014 ref-frame
   expiry or the ADR 0012/0016 staged protocols. Distinct from the same-response settled
@@ -321,7 +322,7 @@ The perfect-shape refactor is complete and merged. Its end-state:
   theirs.
 - Type-cycle growth (R9). R4 keeps the VALUE import graph acyclic, so every remaining cycle is
   created by type-only imports — free at runtime, invisible to R5/R6, and the largest single
-  obstacle to reading a subsystem in isolation: inside a strongly-connected component of 77 files,
+  obstacle to reading a subsystem in isolation: inside a strongly-connected component of 76 files,
   no file has a self-contained slice. `TYPE_CYCLE_BASELINE`, derived from the zone ceilings in
   `scripts/layering/daemon-modularity.ts`, ratchets it for **growth only**, deliberately unlike R6: reducing it
   is a real refactor rather than a file move, so a hard equality would turn every unrelated
@@ -329,7 +330,7 @@ The perfect-shape refactor is complete and merged. Its end-state:
   failing. Hubs by in-component dependents: `runtime-contract.ts` (25),
   `commands/runtime-types.ts` (21), `backend.ts` (15), `commands/runtime-common.ts` (12).
 - Daemon modularity ratchets (R10). The same tooling-only declaration pins R7's writer-owned
-  field/owner-claim counts, R9's 77 members by zone (`commands` 33, `daemon-server` 21, `core` 10,
+  field/owner-claim counts, R9's 76 members by zone (`commands` 33, `daemon-server` 20, `core` 10,
   `platforms` 7, root 5, `client` 1), and the external production importers of `daemon/types.ts`
   (down to 2: the client normalizers and remote artifacts). R7 counts and external importers may
   only shrink; no zone may grow inside R9, and replay/Maestro/replay-test engine files remain outside
